Refactor test cases to use TestSuite::doTest pattern
- Move port_sharing_test.cxx from tests/integration/ to tests/asio/
- Refactor port_sharing_test.cxx to use TestSm/TestMgr from raft_functional_common
- Refactor port_sharing_test.cxx to use TestSuite::doTest instead of custom main()
- Replace std::cout/std::cerr with TestSuite::_msg in all test files
- Simplify dispatcher_test.cxx by removing unnecessary namespace wrapper
- Update tests/CMakeLists.txt to reflect new port_sharing_test location
- Remove tests/integration directory
