+++ tests/unit/launcher_test.cxx
@@ -16,7 +16,7 @@ limitations under the License.
 
 #include "launcher.hxx"
 #include "logger.hxx"
-#include <iostream>
+#include "test_common.h"
 #include <string>
 #include <memory>
 #include <stdexcept>
@@ -47,31 +47,25 @@ public:
 using namespace nuraft;
 
 // Test functions
-void test_init_shared_port_basic() {
+static int test_init_shared_port_basic() {
     raft_launcher launcher;
     ptr<logger> logger = cs_new<simple_logger>();
     asio_service::options asio_opts;
 
     bool result = launcher.init_shared_port(10001, logger, asio_opts);
-
-    if (!result) {
-        throw std::runtime_error("init_shared_port failed");
-    }
+    CHK_TRUE(result);
 
     ptr<asio_service> asio_svc = launcher.get_asio_service();
     ptr<rpc_listener> listener = launcher.get_rpc_listener();
 
-    if (!asio_svc) {
-        throw std::runtime_error("asio_service is null");
-    }
-    if (!listener) {
-        throw std::runtime_error("rpc_listener is null");
-    }
+    CHK_NONNULL(asio_svc);
+    CHK_NONNULL(listener);
 
     launcher.shutdown(1);
+    return 0;
 }
 
-void test_get_raft_server_null_in_shared_mode() {
+static int test_get_raft_server_null_in_shared_mode() {
     // Test that get_raft_server returns null in shared port mode
     raft_launcher launcher;
     ptr<logger> logger = cs_new<simple_logger>();
@@ -80,14 +74,13 @@ void test_get_raft_server_null_in_shared_mode() {
     launcher.init_shared_port(10002, logger, asio_opts);
 
     ptr<raft_server> legacy_server = launcher.get_raft_server();
-    if (legacy_server != nullptr) {
-        throw std::runtime_error("get_raft_server should return null in shared port mode");
-    }
+    CHK_NULL(legacy_server);
 
     launcher.shutdown(1);
+    return 0;
 }
 
-void test_double_init_fails() {
+static int test_double_init_fails() {
     // Test that double initialization fails
     raft_launcher launcher;
     ptr<logger> logger = cs_new<simple_logger>();
@@ -95,20 +88,17 @@ void test_double_init_fails() {
 
     // First init
     bool result = launcher.init_shared_port(10003, logger, asio_opts);
-    if (!result) {
-        throw std::runtime_error("first init_shared_port failed");
-    }
+    CHK_TRUE(result);
 
     // Second init should fail
     result = launcher.init_shared_port(10004, logger, asio_opts);
-    if (result) {
-        throw std::runtime_error("second init_shared_port should fail");
-    }
+    CHK_TRUE(!result);
 
     launcher.shutdown(1);
+    return 0;
 }
 
-void test_get_server_null_before_add() {
+static int test_get_server_null_before_add() {
     // Test that get_server returns null before adding any groups
     raft_launcher launcher;
     ptr<logger> logger = cs_new<simple_logger>();
@@ -117,14 +107,13 @@ void test_get_server_null_before_add() {
     launcher.init_shared_port(10005, logger, asio_opts);
 
     ptr<raft_server> server = launcher.get_server(1);
-    if (server != nullptr) {
-        throw std::runtime_error("get_server should return null before add_group");
-    }
+    CHK_NULL(server);
 
     launcher.shutdown(1);
+    return 0;
 }
 
-void test_remove_nonexistent_group() {
+static int test_remove_nonexistent_group() {
     // Test that removing non-existent group fails
     raft_launcher launcher;
     ptr<logger> logger = cs_new<simple_logger>();
@@ -133,14 +122,13 @@ void test_remove_nonexistent_group() {
     launcher.init_shared_port(10006, logger, asio_opts);
 
     int ret = launcher.remove_group(999);
-    if (ret != -1) {
-        throw std::runtime_error("remove_group for non-existent group should fail");
-    }
+    CHK_EQ(-1, ret);
 
     launcher.shutdown(1);
+    return 0;
 }
 
-void test_add_group_without_init_fails() {
+static int test_add_group_without_init_fails() {
     // Test that add_group fails without init_shared_port
     raft_launcher launcher;
     ptr<logger> logger = cs_new<simple_logger>();
@@ -154,36 +142,19 @@ void test_add_group_without_init_fails() {
     ptr<state_mgr> smgr = nullptr;
 
     int ret = launcher.add_group(1, sm, smgr, params);
-    if (ret != -1) {
-        throw std::runtime_error("add_group should fail without init_shared_port");
-    }
+    CHK_EQ(-1, ret);
+    return 0;
 }
 
-int main() {
-    std::cout << "Running launcher tests..." << std::endl;
-    try {
-        test_init_shared_port_basic();
-        std::cout << "  ✓ init_shared_port_basic" << std::endl;
+int main(int argc, char** argv) {
+    TestSuite ts;
 
-        test_get_raft_server_null_in_shared_mode();
-        std::cout << "  ✓ get_raft_server_null_in_shared_mode" << std::endl;
+    ts.doTest("init_shared_port basic", test_init_shared_port_basic);
+    ts.doTest("get_raft_server null in shared mode", test_get_raft_server_null_in_shared_mode);
+    ts.doTest("double init fails", test_double_init_fails);
+    ts.doTest("get_server null before add", test_get_server_null_before_add);
+    ts.doTest("remove nonexistent group", test_remove_nonexistent_group);
+    ts.doTest("add_group without init fails", test_add_group_without_init_fails);
 
-        test_double_init_fails();
-        std::cout << "  ✓ double_init_fails" << std::endl;
-
-        test_get_server_null_before_add();
-        std::cout << "  ✓ get_server_null_before_add" << std::endl;
-
-        test_remove_nonexistent_group();
-        std::cout << "  ✓ remove_nonexistent_group" << std::endl;
-
-        test_add_group_without_init_fails();
-        std::cout << "  ✓ add_group_without_init_fails" << std::endl;
-
-        std::cout << "\nAll tests passed!" << std::endl;
-        return 0;
-    } catch (const std::exception& e) {
-        std::cerr << "\nTest failed: " << e.what() << std::endl;
-        return 1;
-    }
+    return 0;
 }
